+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 - **[Feature]** Add `ErasedScheduler` trait, a dyn-compatible version of `Scheduler`.
+- **[Feature]** Add alias trait for scheduler based on `StdClock` and `ChronoClock`.
 
 # 0.1.0 (2025-06-04)